API: Adds JWT Auth

+++ API/src/main.cr
@@ -1,20 +1,14 @@
 require "kemal"
 require "sqlite3"
+require "crypto/bcrypt"
+require "jwt"
 
-# start CORS
+# enable cors
+require "./cors"
 
-before_all do |env|
-  env.response.headers["Access-Control-Allow-Origin"] = "*"
-end
-
-options "/*" do |env|
-  env.response.headers["Access-Control-Allow-Methods"] = "HEAD,GET,PUT,POST,DELETE,OPTIONS"
-  env.response.headers["Access-Control-Allow-Headers"] = "X-Requested-With, X-HTTP-Method-Override, Content-Type, Cache-Control, Accept"
-
-  halt env, 200
-end
-
-# end CORS
+# auth middleware
+require "./auth_handler"
+add_handler AuthHandler.new
 
 get "/" do
   "Journals API"
@@ -25,13 +19,25 @@ db.exec("PRAGMA foreign_keys = ON")
 
 get "/install" do
   db.exec "
+    CREATE TABLE IF NOT EXISTS users (
+        id INTEGER,
+        username TEXT,
+        password TEXT,
+        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+        PRIMARY KEY(id)
+    );
+  "
+  db.exec "
     CREATE TABLE IF NOT EXISTS notebooks (
         id INTEGER,
         name TEXT,
         expanded INTEGER DEFAULT 1,
+        user_id INTEGER,
         created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
         updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        PRIMARY KEY(id)
+        PRIMARY KEY(id),
+        FOREIGN KEY(user_id) REFERENCES users(id) ON DELETE CASCADE
     );
   "
   db.exec "
@@ -39,10 +45,12 @@ get "/install" do
         id INTEGER,
         notebook_id INTEGER,
         name TEXT,
+        user_id INTEGER,
         created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
         updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
         PRIMARY KEY(id),
         FOREIGN KEY(notebook_id) REFERENCES notebooks(id) ON DELETE CASCADE
+        FOREIGN KEY(user_id) REFERENCES users(id) ON DELETE CASCADE
     );
   "
   db.exec "
@@ -52,17 +60,84 @@ get "/install" do
         name TEXT,
         type TEXT,
         content TEXT,
+        user_id INTEGER,
         created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
         updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
         PRIMARY KEY(id),
         FOREIGN KEY(section_id) REFERENCES sections(id) ON DELETE CASCADE
+        FOREIGN KEY(user_id) REFERENCES users(id) ON DELETE CASCADE
     );
   "
   "Installation Complete!"
 end
 
+post "/register" do |env|
+  username = env.params.json["username"]?
+  password = env.params.json["password"]?
+
+  env.response.content_type = "application/json"
+
+  if !username || !password
+    env.response << {error: "username or password not provided"}.to_json
+    next
+  end
+
+  username = username.as(String)
+  password = password.as(String)
+
+  user = db.query_one?("SELECT username FROM users WHERE username = ?", username, as: {
+    username: String,
+  })
+
+  if user
+    env.response << {"error": "User already exists"}.to_json
+    next
+  end
+
+  hashed_password = Crypto::Bcrypt::Password.create(password).to_s
+  result = db.exec "INSERT INTO users(username, password) VALUES(?, ?)", username, hashed_password
+
+  {success: true}.to_json
+end
+
+post "/login" do |env|
+  username = env.params.json["username"]?
+  password = env.params.json["password"]?
+
+  env.response.content_type = "application/json"
+
+  if !username || !password
+    env.response << {error: "username or password not provided"}.to_json
+    next
+  end
+
+  username = username.as(String)
+  password = password.as(String)
+
+  user = db.query_one?("SELECT username, password FROM users WHERE username = ?", username, as: {
+    username: String,
+    password: String,
+  })
+
+  env.response.content_type = "application/json"
+
+  if user
+    stored_password = Crypto::Bcrypt::Password.new(user["password"])
+    if stored_password.verify(password)
+      exp = Time.now.to_unix + (60 * 5)
+      payload = {username: user["username"], exp: exp}
+      jwt = JWT.encode(payload, "12345", JWT::Algorithm::HS256)
+      {token: jwt, expiresIn: "5 minutes"}.to_json
+    else
+      {error: "Authentication Failed"}.to_json
+    end
+  else
+    {error: "User not found"}.to_json
+  end
+end
+
 get "/notebooks" do |env|
-  notebooks = db.query_all("SELECT * from notebooks", as: {
+  notebooks = db.query_all("SELECT * from notebooks WHERE user_id = ?", env.auth_id, as: {
     id:         Int64,
     name:       String,
     expanded:   Bool,
@@ -92,7 +167,7 @@ end
 
 post "/notebooks" do |env|
   notebook_name = env.params.json["notebookName"].as(String)
-  result = db.exec "INSERT INTO notebooks(name) VALUES(?)", notebook_name
+  result = db.exec "INSERT INTO notebooks(name, user_id) VALUES(?, ?)", notebook_name, env.auth_id
 
   env.response.content_type = "application/json"
   {insertedRowId: result.last_insert_id}.to_json
@@ -101,7 +176,7 @@ end
 post "/sections" do |env|
   notebook_id = env.params.json["notebookId"].as(Int64)
   section_name = env.params.json["sectionName"].as(String)
-  result = db.exec "INSERT INTO sections(notebook_id, name) VALUES(?, ?)", notebook_id, section_name
+  result = db.exec "INSERT INTO sections(notebook_id, name, user_id) VALUES(?, ?, ?)", notebook_id, section_name, env.auth_id
 
   env.response.content_type = "application/json"
   {insertedRowId: result.last_insert_id}.to_json
@@ -111,7 +186,7 @@ post "/pages" do |env|
   section_id = env.params.json["sectionId"].as(Int64)
   page_type = env.params.json["pageType"].as(String)
   page_name = env.params.json["pageName"].as(String)
-  result = db.exec "INSERT INTO pages(section_id, type, name) VALUES(?, ?, ?)", section_id, page_type, page_name
+  result = db.exec "INSERT INTO pages(section_id, type, name, user_id) VALUES(?, ?, ?, ?)", section_id, page_type, page_name, env.auth_id
 
   env.response.content_type = "application/json"
   {insertedRowId: result.last_insert_id}.to_json
@@ -120,7 +195,7 @@ end
 get "/pages/:section_id" do |env|
   section_id = env.params.url["section_id"]
 
-  pages = db.query_all("SELECT id, name, type from pages WHERE section_id = ?", section_id, as: {
+  pages = db.query_all("SELECT id, name, type from pages WHERE section_id = ? AND user_id = ?", section_id, env.auth_id, as: {
     id:   Int64,
     name: String,
     type: String,
@@ -132,7 +207,7 @@ end
 
 delete "/pages/:page_id" do |env|
   page_id = env.params.url["page_id"]
-  db.exec "DELETE FROM pages WHERE id = ?", page_id
+  db.exec "DELETE FROM pages WHERE id = ? AND user_id = ?", page_id, env.auth_id
 
   env.response.content_type = "application/json"
   {success: true}.to_json
@@ -140,7 +215,7 @@ end
 
 delete "/sections/:section_id" do |env|
   section_id = env.params.url["section_id"]
-  db.exec "DELETE FROM sections WHERE id = ?", section_id
+  db.exec "DELETE FROM sections WHERE id = ? AND user_id = ?", section_id, env.auth_id
 
   env.response.content_type = "application/json"
   {success: true}.to_json
@@ -148,7 +223,7 @@ end
 
 delete "/notebooks/:notebook_id" do |env|
   notebook_id = env.params.url["notebook_id"]
-  db.exec "DELETE FROM notebooks WHERE id = ?", notebook_id
+  db.exec "DELETE FROM notebooks WHERE id = ? AND user_id = ?", notebook_id, env.auth_id
 
   env.response.content_type = "application/json"
   {success: true}.to_json
@@ -158,7 +233,7 @@ put "/pages/:page_id" do |env|
   page_id = env.params.url["page_id"]
   page_content = env.params.json["pageContent"].as(String)
 
-  db.exec "UPDATE pages SET content=? WHERE id = ?", page_content, page_id
+  db.exec "UPDATE pages SET content=? WHERE id = ? AND user_id = ?", page_content, page_id, env.auth_id
 
   env.response.content_type = "application/json"
   {success: true}.to_json
@@ -167,7 +242,7 @@ end
 get "/pages/content/:page_id" do |env|
   page_id = env.params.url["page_id"]
 
-  page = db.query_one("SELECT content from pages WHERE id = ?", page_id, as: {content: Nil | String})
+  page = db.query_one("SELECT content from pages WHERE id = ? AND user_id = ?", page_id, env.auth_id, as: {content: Nil | String})
 
   env.response.content_type = "application/json"
   page.to_json
@@ -177,7 +252,7 @@ put "/pages/name/:page_id" do |env|
   page_id = env.params.url["page_id"]
   page_name = env.params.json["pageName"].as(String)
 
-  db.exec "UPDATE pages SET name=? WHERE id = ?", page_name, page_id
+  db.exec "UPDATE pages SET name=? WHERE id = ? AND user_id = ?", page_name, page_id, env.auth_id
 
   env.response.content_type = "application/json"
   {success: true}.to_json
@@ -187,7 +262,7 @@ put "/sections/name/:section_id" do |env|
   section_id = env.params.url["section_id"]
   section_name = env.params.json["sectionName"].as(String)
 
-  db.exec "UPDATE sections SET name=? WHERE id = ?", section_name, section_id
+  db.exec "UPDATE sections SET name=? WHERE id = ? AND user_id = ?", section_name, section_id, env.auth_id
 
   env.response.content_type = "application/json"
   {success: true}.to_json
@@ -197,7 +272,7 @@ put "/notebooks/name/:notebook_id" do |env|
   notebook_id = env.params.url["notebook_id"]
   notebook_name = env.params.json["notebookName"].as(String)
 
-  db.exec "UPDATE notebooks SET name=? WHERE id = ?", notebook_name, notebook_id
+  db.exec "UPDATE notebooks SET name=? WHERE id = ? AND user_id = ?", notebook_name, notebook_id, env.auth_id
 
   env.response.content_type = "application/json"
   {success: true}.to_json
